requirements.txt
google-generativeai>=0.3.0
psycopg2-binary>=2.9.9
testcontainers>=3.7.0
python-dotenv>=0.19.0  # Optional for environment variables